unsorted-guides/stream-processing-fundamentals/ui/requirements.txt
click==8.1.3
dash==2.8.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
hazelcast-python-client==5.3.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
pandas==1.3.5
plotly==5.13.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
six==1.16.0
tenacity==8.1.0
Werkzeug==2.2.2
zipp==3.11.0
